+++ packages/ato/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @ATO/ATO
+# @maximdevoir/ato
 
 ## 1.0.1
 
@@ -7,7 +7,7 @@
 - [`f3a01bd`](https://github.com/MaximDevoir/ATO/commit/f3a01bdf9c5a7016695f512f65081ff50f448f56) Thanks [@MaximDevoir](https://github.com/MaximDevoir)! - fix: add release script
 
 - Updated dependencies [[`f3a01bd`](https://github.com/MaximDevoir/ATO/commit/f3a01bdf9c5a7016695f512f65081ff50f448f56)]:
-  - @ATO/UnrealLag@1.0.1
+  - @maximdevoir/unreal-lag@1.0.1
 
 ## 1.0.0
 
@@ -22,4 +22,4 @@
 ### Patch Changes
 
 - Updated dependencies [[`d85bfcf`](https://github.com/MaximDevoir/ATO/commit/d85bfcf4ae7ebc995bb659a02eb45bfed52e4f6f), [`a2b0466`](https://github.com/MaximDevoir/ATO/commit/a2b0466c02d4cec2516de77f38c215cc170b69b3)]:
-  - @ATO/UnrealLag@1.0.0
+  - @maximdevoir/unreal-lag@1.0.0